#355 rearrange tests, make system plan tests
