+++ src/artemis/unit_tests/test_log.py
